+++ CMakeLists.txt
@@ -208,7 +208,7 @@ add_library(rdf4cpp
         src/rdf4cpp/rdf/storage/tuple/DefaultSolutionSequenceBackend.cpp
         src/rdf4cpp/rdf/storage/tuple/IDatasetBackend.cpp
         src/rdf4cpp/rdf/storage/tuple/ISolutionSequenceBackend.cpp
-        src/rdf4cpp/rdf/writer/Serialization.cpp
+        src/rdf4cpp/rdf/writer/SerializationState.cpp
         src/rdf4cpp/rdf/IRIView.cpp
         src/rdf4cpp/rdf/IRIFactory.cpp
         private/rdf4cpp/rdf/parser/IStreamQuadIteratorSerdImpl.cpp
