--- src/rdf4cpp/rdf/writer/WriterBasics.hpp
@@ -1,8 +1,0 @@
-#ifndef RDF4CPP_WRITERBASICS_HPP
-#define RDF4CPP_WRITERBASICS_HPP
-
-namespace rdf4cpp::rdf::writer {
-class SerializationState;
-}
-
-#endif  //RDF4CPP_WRITERBASICS_HPP
